requirements.txt
PyPDF2
Pillow
customtkinter
tabula-py
